requirements.txt
flask>=2.3.0
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
MarkupSafe>=2.1.0
